requirements.txt
# ============================================================
# 🚛 Shipment Delay Early Warning System — Dependencies
# ============================================================

# ── Core Framework ───────────────────────────────────────
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Database ─────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Authentication ───────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# ── ML / Data Science ───────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
shap>=0.43.0

# ── Visualization ───────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.13.0

# ── HTTP Client (for real weather API) ──────────────────
httpx>=0.25.0

# ── Caching (optional) ─────────────────────────────────
redis>=5.0.0

# ── Testing ─────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
